% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ARBinProbitBART.R
\name{ARProbitbartNOCovars_fullcond}
\alias{ARProbitbartNOCovars_fullcond}
\title{Auto-regressive Probit BART Model without exogenous covariates}
\usage{
ARProbitbartNOCovars_fullcond(
  y.train = NULL,
  n.iter = 1000,
  n.burnin = 100,
  num_test_periods = 1,
  num_lags = 1,
  indiv.offsets = FALSE,
  num_z_iters = 10,
  noise_in_pred = 1
)
}
\arguments{
\item{y.train}{A vector (or matrix for panel data) of binary outcomes. Rows correspond to time periods, columns correspond to variables.}

\item{n.iter}{Number of iterations excluding burnin.}

\item{n.burnin}{Number of burnin iterations.#'}

\item{num_test_periods}{Number of future time periods to predict.}

\item{num_lags}{Number of lags of latent outcome to be included as potential splitting variables. Currently must be equal to 1.}

\item{indiv.offsets}{Include individual-specific offsets? Only relevant if yttrain contains more than one column. Currently not supported.}

\item{num_z_iters}{Number of Gibbs iterations for latent outcome z samples per overall MCMC iteration.}

\item{noise_in_pred}{If equal to 1, keep noise in test prediction calculations.}
}
\value{
A list is returned containing the following elements:
\item{ydraws_train}{Samples from the posterior of binary outcomes for training observations.}
\item{ydraws_train}{Samples from the posterior of binary outcomes for test observations}
}
\description{
Auto-regressive (of order one) Probit BART Model without exogenous covariates
}
